### 2019-06-06 #1492 @whedon
[PRE REVIEW]: STUMPY: A Powerful and Scalable Python Library for Time Series Data Mining
**Submitting author:** @seanmylaw (<a href="http://orcid.org/0000-0003-2594-1969">Sean Law</a>)
**Repository:** <a href="https://github.com/TDAmeritrade/stumpy" target ="_blank">https://github.com/TDAmeritrade/stumpy</a>
**Version:** v1.0
