### 2019-06-06 2c1a9ac
Changed tasks to unordered bulleted list
