### 2019-06-07 0946f11
Split up array tests from dataframe/series tests
