+++ tests/test_stumped.py
@@ -80,7 +80,6 @@ test_data = [
 @pytest.mark.filterwarnings("ignore:\\s+Port 8787 is already in use:UserWarning")
 @pytest.mark.parametrize("T_A, T_B", test_data)
 def test_stumped_self_join(T_A, T_B, dask_client):
-    dask_client.restart()
     m = 3
     zone = int(np.ceil(m / 4))
     left = np.array(
@@ -94,12 +93,27 @@ def test_stumped_self_join(T_A, T_B, dask_client):
     replace_inf(left)
     replace_inf(right)
     npt.assert_almost_equal(left, right)
-    dask_client.restart()
 
+
+@pytest.mark.filterwarnings("ignore:numpy.dtype size changed")
+@pytest.mark.filterwarnings("ignore:numpy.ufunc size changed")
+@pytest.mark.filterwarnings("ignore:numpy.ndarray size changed")
+@pytest.mark.filterwarnings("ignore:\\s+Port 8787 is already in use:UserWarning")
+@pytest.mark.parametrize("T_A, T_B", test_data)
+def test_stumped_self_join_df(T_A, T_B, dask_client):
+    m = 3
+    zone = int(np.ceil(m / 4))
+    left = np.array(
+        [
+            naive_mass(Q, T_B, m, i, zone, True)
+            for i, Q in enumerate(core.rolling_window(T_B, m))
+        ],
+        dtype=object,
+    )
     right = stumped(dask_client, pd.Series(T_B), m, ignore_trivial=True)
+    replace_inf(left)
     replace_inf(right)
     npt.assert_almost_equal(left, right)
-    dask_client.restart()
 
 
 @pytest.mark.filterwarnings("ignore:numpy.dtype size changed")
@@ -108,7 +122,6 @@ def test_stumped_self_join(T_A, T_B, dask_client):
 @pytest.mark.filterwarnings("ignore:\\s+Port 8787 is already in use:UserWarning")
 @pytest.mark.parametrize("T_A, T_B", test_data)
 def test_stumped_A_B_join(T_A, T_B, dask_client):
-    dask_client.restart()
     m = 3
     left = np.array(
         [naive_mass(Q, T_A, m) for Q in core.rolling_window(T_B, m)], dtype=object
@@ -117,11 +130,21 @@ def test_stumped_A_B_join(T_A, T_B, dask_client):
     replace_inf(left)
     replace_inf(right)
     npt.assert_almost_equal(left, right)
-    dask_client.restart()
 
+
+@pytest.mark.filterwarnings("ignore:numpy.dtype size changed")
+@pytest.mark.filterwarnings("ignore:numpy.ufunc size changed")
+@pytest.mark.filterwarnings("ignore:numpy.ndarray size changed")
+@pytest.mark.filterwarnings("ignore:\\s+Port 8787 is already in use:UserWarning")
+@pytest.mark.parametrize("T_A, T_B", test_data)
+def test_stumped_A_B_join_df(T_A, T_B, dask_client):
+    m = 3
+    left = np.array(
+        [naive_mass(Q, T_A, m) for Q in core.rolling_window(T_B, m)], dtype=object
+    )
     right = stumped(
         dask_client, pd.Series(T_A), m, pd.Series(T_B), ignore_trivial=False
     )
+    replace_inf(left)
     replace_inf(right)
     npt.assert_almost_equal(left, right)
-    dask_client.restart()
