### 2019-06-09 4e7ea36
Changed tutorial URLs
