### 2019-06-09 ed98051
Fixed incorrect references on RTD
