### 2019-06-09 #1492 @labarba
@seanlaw — I'm going to send this back to you for some more work, because the paper itself does not quite meet [our expectations](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). Right now, the article is very brief and doesn't communicate a lot about the software, and its need or uses in research. While we don't expect a full article, it should be 250–1000 words (currently you have ~130 words). These items need to be improved:

- the summary describing high-level functionality and purpose of the software for a diverse, non-specialist audience;
- a clear statement of need that illustrates the research application of the software (including references to other software meeting similar needs).

In essence, the article needs to be improved with more description of how the software works, some examples, and use cases.
