### 2019-06-09 43a0474
Added additional paragraph for JOSS requirements
