### 2019-06-09 #1492 @seanlaw
@labarba thank you for the feedback. I've added a summary that hopefully motivates the need for this software and I've, hopefully, clarified what the core competency (the efficient calculation of the matrix profile - which is the most computationally intensive part of motif (i.e., pattern) and discord (i.e., novelty/anomaly) discovery).

Let me know if or where I can be more clear. Perhaps, you can direct me to some reasonably good examples that I can learn from.
