### 2019-06-10 #1492 @danielskatz
Thanks @seanlaw - this looks good to go forward.  While I find an editor, you might want to fix the references - issues where letters are incorrectly in lower case can be fixed by putting {}s around letters of strings that should not be converted to lower case by bibtex.
