### 2019-06-10 64ecffd
Updated bibtext file to maintain upper case letters
