+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @inproceedings{yeh2016,
-    title={Matrix Profile I: All Pairs Similarity Joins for Time Series: A Unifying View That Includes Motifs, Discords and Shapelets},
+    title={{Matrix Profile I: All Pairs Similarity Joins for Time Series: A Unifying View That Includes Motifs, Discords and Shapelets}},
     author={Yeh, Michael Chin-Chia and Zhu, Yan and Ulanova, Liudmila and Nurjahan, Begum and Ding, Yifei and Dau, Hoang Anh and Silva, Diego Furtado and Mueen, Adullah and Keogh, Eamonn},
-    booktitle={International Conference on Data Mining (ICDM)},
+    booktitle={{International Conference on Data Mining (ICDM)}},
     pages={1317--1322},
     year={2016},
     organization={IEEE},
@@ -10,9 +10,9 @@
 }
 
 @inproceedings{zhu2016,
-    title={Matrix Profile II: Exploiting a Novel Algorithm and GPUs to Break the One Hundred Million Barrier for Time Series Motifs and Joins},
+    title={{Matrix Profile II: Exploiting a Novel Algorithm and GPUs to Break the One Hundred Million Barrier for Time Series Motifs and Joins}},
     author={Zhu, Yan and Zimmerman, Zachary and Senobari, Nader Shakibay and Yeh, Michael Chin-Chia and Funning, Gareth and Mueen, Adullah and Brisk, Philip and Keogh, Eamonn},
-    booktitle={International Conference on Data Mining (ICDM)},
+    booktitle={{International Conference on Data Mining (ICDM)}},
     pages={739--748},
     year={2016},
     organization={IEEE},
@@ -21,9 +21,9 @@
 }
 
 @inproceedings{yeh2017,
-    title={Matrix Profile VI: Meaningful Multidimensional Motif Discovery},
+    title={{Matrix Profile VI: Meaningful Multidimensional Motif Discovery}},
     author={Yeh, Michael Chin-Chia and Kavantzas, Nickolas and Keogh, Eamonn},
-    booktitle={International Conference on Data Mining (ICDM)},
+    booktitle={{International Conference on Data Mining (ICDM)}},
     pages={565--574},
     year={2017},
     organization={IEEE},
@@ -32,9 +32,9 @@
 }
 
 @inproceedings{zhu2017,
-    title={Matrix Profile VII: Time Series Chains: A New Primitive for Time Series Data Mining},
+    title={{Matrix Profile VII: Time Series Chains: A New Primitive for Time Series Data Mining}},
     author={Zhu, Yan and Imamura, Makoto and Nikovski, Daniel and Keogh, Eamonn},
-    booktitle={International Conference on Data Mining (ICDM)},
+    booktitle={{International Conference on Data Mining (ICDM)}},
     pages={695--704},
     year={2017},
     organization={IEEE},
