### 2019-06-10 #1492 @seanlaw
> Thanks @seanlaw - this looks good to go forward. While I find an editor, you might want to fix the references - issues where letters are incorrectly in lower case can be fixed by putting {}s around letters of strings that should not be converted to lower case by bibtex.

@danielskatz Thanks for the pro tip! It should be fixed now.

@VivianePons Best of luck but I'm sure you'll do great!
