### 2019-06-17 7ebba3e
Point binder badge to notebooks directory
