+++ README.rst
@@ -23,7 +23,7 @@
     :target: https://codecov.io/gh/TDAmeritrade/stumpy
     :alt: Code Coverage
 .. image:: https://mybinder.org/badge_logo.svg
-    :target: https://mybinder.org/v2/gh/TDAmeritrade/stumpy/master
+    :target: https://mybinder.org/v2/gh/TDAmeritrade/stumpy/master?filepath=notebooks
     :alt: Binder
 .. image:: https://img.shields.io/twitter/follow/stumpy_dev.svg?style=social
     :target: https://twitter.com/stumpy_dev
