### 2019-06-20 f3a1c8c
Added TDA requirement for language in code of conduct
