### 2019-06-24 07a7396
Add license scan report and status
Signed-off-by: fossabot <badges@fossa.io>
