+++ README.rst
@@ -1,5 +1,6 @@
-[![FOSSA Status](https://app.fossa.io/api/projects/git%2Bgithub.com%2FTDAmeritrade%2Fstumpy.svg?type=shield)](https://app.fossa.io/projects/git%2Bgithub.com%2FTDAmeritrade%2Fstumpy?ref=badge_shield)
-
+.. image:: https://app.fossa.io/api/projects/git%2Bgithub.com%2FTDAmeritrade%2Fstumpy.svg?type=shield
+    :target: https://app.fossa.io/projects/git%2Bgithub.com%2FTDAmeritrade%2Fstumpy?ref=badge_shield
+    :alt: FOSSA Status
 .. image:: https://img.shields.io/pypi/v/stumpy.svg
     :target: https://pypi.org/project/stumpy/
     :alt: PyPI Version
@@ -283,5 +284,6 @@ License & Trademark
 | Copyright 2019 TD Ameritrade. Released under the terms of the 3-Clause BSD license.
 | STUMPY is a trademark of TD Ameritrade IP Company, Inc. All rights reserved.
 
-
-[![FOSSA Status](https://app.fossa.io/api/projects/git%2Bgithub.com%2FTDAmeritrade%2Fstumpy.svg?type=large)](https://app.fossa.io/projects/git%2Bgithub.com%2FTDAmeritrade%2Fstumpy?ref=badge_large)
+.. image:: https://app.fossa.io/api/projects/git%2Bgithub.com%2FTDAmeritrade%2Fstumpy.svg?type=large
+    :target: https://app.fossa.io/projects/git%2Bgithub.com%2FTDAmeritrade%2Fstumpy?ref=badge_large
+    :alt: FOSSA Status
