+++ docs/motivation.rst
@@ -1,0 +1,11 @@
+-----------------------
+Background & Motivation
+-----------------------
+
+The following video from May 2019, provides the background and motivation for developing and open sourcing STUMPY:
+
+.. raw:: html
+
+    <div style="position: relative; padding-bottom: 56.25%; height: 0; overflow: hidden; max-width: 100%; height: auto;">
+        <iframe src="https://www.youtube.com/embed/ep6yBul2Sq8" frameborder="0" allowfullscreen style="position: absolute; top: 0; left: 0; width: 100%; height: 100%;"></iframe>
+    </div>
