### 2019-07-07 b44875a
Fixed #45 Window Size <= 2
