+++ stumpy/stomp.py
@@ -65,6 +65,7 @@ def stomp(T_A, m, T_B=None, ignore_trivial=True):
     if T_B is None:
         T_B = T_A
     core.check_dtype(T_B)
+    core.check_window_size(m)
 
     if ignore_trivial is False and core.are_arrays_equal(T_A, T_B):  # pragma: no cover
         logger.warning("Arrays T_A, T_B are equal, which implies a self-join.")
