+++ stumpy/mstump.py
@@ -421,6 +421,7 @@ def mstump(T, m):
 
     T = core.df_to_array(T)
     core.check_dtype(T)
+    core.check_window_size(m)
 
     d = T.shape[0]
     n = T.shape[1]
