### 2019-07-15 TDAmeritrade/stumpy#50 @seanlaw
Typo in Times Series Chains Example in README
The time series chains example shows:

    left_matrix_profile_index = matrix_profile[2]
    right_matrix_profile_index = matrix_profile[3]

and, instead, should say:

    left_matrix_profile_index = matrix_profile[:, 2]
    right_matrix_profile_index = matrix_profile[:, 3]
