### 2019-07-15 d4f0839
Fixed #50 Typo in Time Series Chains Example
