### 2019-07-15 #1504 @seanlaw
@hooman650 Thank you for your thorough review.

Regarding 2:
> The author has prepared a nice documentation page as well as contributing guidelines and examples. However, I just ran into an exception as I tried to run the following example from the documentation:

Indeed, this was a typo and we have submitted an [issue](https://github.com/TDAmeritrade/stumpy/issues/50) and [fixed it accordingly](https://github.com/TDAmeritrade/stumpy/commit/d4f08395eef34e61022a30f51f59730839f762bb) in both our README and ReadTheDocs. Thank you for pointing this out!

Regarding 3:
> A comment in regards to the performance comparison graph that is mentioned here; How did the authors of GPU-STOMP implement their algorithm? In tensorflow? I feel like a good implementation is totally feasible in tensorflow, given that recently almost every hard computation is performing way better on GPUs compared to CPUs, I have difficulties believing a good GPU implementation would be beaten by CPUs. But, I might be wrong.

Unfortunately, I am not the original author of the GPU-STOMP publication/code and the numbers shown were simply extracted from their published paper for comparison. Currently, I only have access to CPUs but it is a top priority in our [project roadmap](https://github.com/TDAmeritrade/stumpy/issues/1) to port this work over to GPUs. We are in the process of looking for assistance and resources from folks at NVIDIA. The initial goal of our scalable CPU implementation was to allow non-tech-savvy scientists to be able to get up and running quickly without needing access to any specialized hardware. We believe that we have achieved this goal.

Regarding 4:
> My final comment is in regards to the "window size" input of STUMPY. This could be compared to kernel size in CNNs. This will be always challenging to know prior to analysis. Any suggestions for determining that?

This is an excellent question and has been discussed by the original authors (not me) in _Section D_ of their paper [matrix profile II](https://www.cs.ucr.edu/~eamonn/STOMP_GPU_final_submission_camera_ready.pdf). In summary, the window size is certainly a _user input_ that requires some level of domain expertise. However, the original authors have demonstrated that the matrix profile is robust to varying window sizes and that being "in the ballpark" is often enough to find motifs.

Regarding 5:
> There is no version release number in the repository.

We currently provide a version number in the standard `setup.py` file and the version number is also accessible within Python via `stumpy.__version__`. Perhaps there is a better or more standard place to specify the version release number? Any guidance would be greatly appreciated.
