### 2019-07-15 #1504 @mbobra
👋 @seanlaw

Github is showing that there [aren't any releases in the repository](https://github.com/TDAmeritrade/stumpy/releases). Could you please go ahead and [create a release](https://help.github.com/en/articles/creating-releases)?

Could you also include your responses to points 3 and 4 above in the text of the paper along with the appropriate references?
