### 2019-07-17 #1504 @ejolly
Hi @mbobra sorry for the delay on my review!

First of all, I'd like to note that Stumpy is a great addition to the time-series analyst's toolkit and is very well-documented, explained, and referenced. I also rather enjoyed the talk. Really nice @seanlaw!

My testing was done using Python 3.6 on Mac OS 10.14.2 and everything installed without issues. I updated my installation given the most recent changes made as per @hooman650's review and can attest that fix for comment 2 now works.

I was unable to test out the performance claims given limited current access to a distributed compute system at this time, but I was able to at least test the functionality by running a local dask server without issues.

Just a few minor suggestions:

1. Since the tutorial examples are provided as jupyter notebooks and the prerendered file for `Tutorial_1` includes inline plots, it would good to add the `%matplotlib inline` command to the top code cell of that notebook so that anyone just downloading and running the notebooks immediately reproduces the prerendered files on github
2. In its present form, `Tutorial_2` is a bit lacking in context and explanation. While it's nice to see how to use some additional functionality and be provided links to the original paper, I might suggest at least a bit more of an explanation as to what and why one might use time-series chains. `Tutorial_1` for example, does a great job in providing a high-level overview of the matrix profile and elucidating the impact of the window size free parameter
3. More of a minor suggestion, but it might be nice to have the _text_ link to the documentation a bit higher up on the github README page, for example in the website field of the repo description or edit the text in the first paragraph to clearly mention the documentation site. While badge and "matrix normal" links work, the actualy documentation site itself isn't discussed until halfway through the README. Feel free to ignore this if you prefer.

With those minor changes I think this would make a great addition to JOSS.
