### 2019-07-17 TDAmeritrade/stumpy#51 @seanlaw
Add Inline Plotting to Tutorial 1
%matplotlib inline

Needs to be added to Tutorial 1 to allow plots to display pre-rendered on Github
