### 2019-07-17 cab4137
Fixed #51 Display Pre-rendered Plots on Github
