### 2019-07-17 4d4e950
Fixed #52 Added link to docs earlier in README
