### 2019-07-17 #1504 @ejolly
@seanlaw the binder addition is a great one and I think will be great for new users.

Regarding my point 2:

I apologize as I should have been more clear. I completely agree with your response that publication in JOSS should be _not_ contingent on you adding a more comprehensive tutorial 2 and my comment was more of a suggestion for something that would improve the tutorials, i.e. would be "nice to have".  From my review, you have already done a fantastic job of documenting, testing, and providing the requisite high-level explanation of the package functionality as per JOSS requirements.

I completely understand that creating and maintaining a solo project is a huge demand on your time and it will be great to see how tutorials and functionality grow with the community base!
