### 2019-07-17 #1504 @seanlaw
@ejolly No need to apologize as I assumed no ill intent. Thank you (@hooman650 and @mbobra) for taking the time to review! I really appreciate the valuable feedback.
