### 2019-07-17 #1504 @mbobra
@hooman650 and @ejolly Thank you so much for reviewing! We really appreciate your time and effort ☀️

@seanlaw We're almost there! Can you please [archive your release on Zenodo](https://guides.github.com/activities/citable-code/) to obtain a DOI and then put that in your README.rst file? After that I think we're done 🎉
