### 2019-07-17 #1504 @seanlaw
@mbobra I've added the DOI as a badge a the top of the README.rst. Is that what you mean?
