+++ README.rst
@@ -26,7 +26,7 @@
     :target: https://mybinder.org/v2/gh/TDAmeritrade/stumpy/master?filepath=notebooks
     :alt: Binder
 .. image:: https://zenodo.org/badge/184809315.svg
-    :target: https://zenodo.org/badge/184809315.svg
+    :target: https://zenodo.org/badge/latestdoi/184809315
     :alt: DOI
 .. image:: https://img.shields.io/twitter/follow/stumpy_dev.svg?style=social
     :target: https://twitter.com/stumpy_dev
