### 2019-07-17 #1504 @mbobra
@openjournals/joss-eics This paper is ready for acceptance! Nice work @seanlaw 🎉
