### 2019-07-18 #1504 @seanlaw
@danielskatz The final proof looks good. Is there anything else that I need to do or is whedon’s command for you to handle? I just want to make sure I am not holding up the process.

<sub>Sent with <a href="http://githawk.com">GitHawk</a></sub>
