### 2019-07-18 #1504 @danielskatz
It's fine - between being on a plane where I couldn't see the final PDF to check it, and then driving and sleeping, I've just gotten back to it :)
