### 2019-07-18 #1504 @danielskatz
Thanks to @ejolly & @hooman650 for reviewing and to @mbobra for editing
