### 2019-08-16 #1504 @seanlaw
@danielskatz I just spotted a minor typo in the PDF. Is there some way that I can fix it?
