### 2019-08-24 #1504 @seanlaw
@arfon I have fixed the typo in the original source repository. Can you please take a look?
